// types/user.ts
export interface User {
  id: number;
  username: string;
  password: string;
  customDomain: string | null;
  address: string | null;
  wechat: string | null;
  email: string | null;
  telegram: string | null;
  twitter: string | null;
  GitHub: string | null;
  bilibili: string | null;
  weibo: string | null;
  QQ: string | null;
  discord: string | null;
  phonenumber: string | null;
  Facebook: string | null;
  ins: string | null;
  steam: string | null;
  douban: string | null;
  xiaohongshu: string | null;
  douyin: string | null;
}